pyproject.toml
[tool.poetry]
name = "test-konfio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
psycopg2-binary = "^2.9.5"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
